chore: update types and unit test
